+++ ext/QuantumOpticsExt/QuantumOpticsExt.jl
@@ -102,6 +102,8 @@ express_nolookup(x::IdentityOp, r::QuantumOpticsRepr) = identityoperator(finite_
 express_nolookup(p::PauliNoiseCPTP, ::QuantumOpticsRepr) = LazySuperSum(SpinBasis(1//2), [1-p.px-p.py-p.pz,p.px,p.py,p.pz],
                                                                [LazyPrePost(_id,_id),LazyPrePost(_x,_x),LazyPrePost(_y,_y),LazyPrePost(_z,_z)])
 
+express_nolookup(s::SOuterKetBra, r::QuantumOpticsRepr) = projector(express(s.ket, r), express(s.bra, r))
+
 include("should_upstream.jl")
 
 end