+++ test/test_misc_linalg.jl
@@ -2,6 +2,12 @@
     @op A; @op B;
     O = SZeroOperator()
 
+    @bra p
+    @bra q
+        
+    @ket m
+    @ket n
+    
     @testset "Complex Conjugate" begin
         @test isequal(conj(O), O)
         @test isequal(conj(conj(A)), A)
@@ -25,4 +31,17 @@
     @testset "Exponential" begin
         @test isequal(exp(A), SExpOperator(A))
     end
+
+    @testset "Vector of Operators" begin
+        @test isequal([1 1;-im im]*[A;B], [A + B;im*B-im*A])  
+    end
+
+    @testset "Vector of Kets" begin
+        @test isequal([1 1;-im im] *[p;q], [p + q;im*q-im*p])
+    end
+
+    @testset "Vector of Bras" begin
+        @test isequal([1 1;-im im] *[m;n], [m + n;im*n-im*m])
+    end
+
 end